V2.8.0

+++ app/lib/screens/logs_screen.dart
@@ -4,6 +4,7 @@ import 'package:menstrudel/database/repositories/periods_repository.dart';
 import 'package:menstrudel/widgets/basic_progress_circle.dart';
 import 'package:menstrudel/models/period_logs/period_day.dart';
 import 'package:menstrudel/models/periods/period.dart';
+import 'package:menstrudel/utils/constants.dart';
 
 import 'package:menstrudel/models/period_prediction_result.dart';
 import 'package:menstrudel/utils/period_predictor.dart';
@@ -13,6 +14,8 @@ import 'package:menstrudel/screens/main_screen.dart';
 import 'package:menstrudel/services/settings_service.dart';
 import 'package:menstrudel/services/period_logger_service.dart';
 import 'package:menstrudel/widgets/logs/dynamic_history_view.dart';
+import 'package:menstrudel/services/wear_sync_service.dart';
+import 'package:menstrudel/widgets/sheets/period_details_bottom_sheet.dart';
 
 import 'package:menstrudel/l10n/app_localizations.dart';
 
@@ -33,19 +36,15 @@ class LogsScreenState extends State<LogsScreen> {
   final periodsRepo = PeriodsRepository();
   final SettingsService _settingsService = SettingsService();
 
+  final _watchSyncService = WatchSyncService();
+
 	List<PeriodDay> _periodLogEntries = [];
   List<Period> _periodEntries = [];
 	bool _isLoading = true;
 	PeriodPredictionResult? _predictionResult;
   PeriodHistoryView _selectedView = PeriodHistoryView.journal;
-
-  Future<void> handleLogPeriod(DateTime selectedDate) async {
-    final bool wasLogSuccessful = await PeriodLoggerService.showAndLogPeriod(context, selectedDate);
-
-    if (wasLogSuccessful && mounted) {
-      _refreshPeriodLogs();
-    }
-  }
+  int _circleCurrentValue = 0;
+  int _circleMaxValue = 28;
   
   Future<void> handleTamponReminder(BuildContext context) async {
     final l10n = AppLocalizations.of(context)!;
@@ -71,13 +70,13 @@ class LogsScreenState extends State<LogsScreen> {
     final l10n = AppLocalizations.of(context)!;
     try {
       await NotificationService.cancelTamponReminder();
-      if (context.mounted) {
+      if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(content: Text(l10n.logScreen_tamponReminderCancelled)),
         );
       }
     } catch (e) {
-      if (context.mounted) {
+      if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(content: Text('${l10n.logScreen_couldNotCancelReminder}: $e'), backgroundColor: Colors.red),
         );
@@ -85,6 +84,23 @@ class LogsScreenState extends State<LogsScreen> {
     } finally {
       _refreshPeriodLogs();
     }
+  }
+
+  void _showDetailsBottomSheet(PeriodDay log) {
+    showModalBottomSheet<void>(
+      context: context,
+      isScrollControlled: true,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
+      ),
+      builder: (context) {
+        return PeriodDetailsBottomSheet(
+          log: log,
+          onDelete: () => _deleteExistingLog(log.id),
+          onSave: _updateExistingLog,
+        );
+      },
+    );
   }
 
 	@override
@@ -102,23 +118,49 @@ class LogsScreenState extends State<LogsScreen> {
 
     if (predictionResult != null) {
       final settingsService = SettingsService();
-      final notificationsEnabled = await settingsService.areNotificationsEnabled();
-      final notificationDays = await settingsService.getNotificationDays();
-      final notificationTime = await settingsService.getNotificationTime();
-      final l10n = AppLocalizations.of(context)!;
-      
-      await NotificationService.schedulePeriodNotification(
-        scheduledTime: predictionResult.estimatedStartDate,
-        areEnabled: notificationsEnabled,
-        daysBefore: notificationDays,
-        notificationTime: notificationTime,
-        title: l10n.notification_periodTitle,
-        body: l10n.notification_periodBody(notificationDays),
-      );
+      final periodNotificationsEnabled = await settingsService.areNotificationsEnabled();
+      final periodNotificationDays = await settingsService.getNotificationDays();
+      final periodNotificationTime = await settingsService.getNotificationTime();
+
+      final periodOverdueNotificationsEnabled = await settingsService.areNotificationsEnabled();
+      final periodOverdueNotificationDays = await settingsService.getNotificationDays();
+      final periodOverdueNotificationTime = await settingsService.getNotificationTime();
+
+      if (mounted){
+        final l10n = AppLocalizations.of(context)!;
+
+        // Period due notification
+        try {
+          await NotificationService.schedulePeriodNotifications(
+            scheduledTime: predictionResult.estimatedStartDate,
+            areEnabled: periodNotificationsEnabled,
+            daysBefore: periodNotificationDays,
+            notificationTime: periodNotificationTime,
+            title: l10n.notification_periodTitle,
+            body: l10n.notification_periodBody(periodNotificationDays),
+            notificationID: periodDueNotificationId,
+          );
+        } catch (e) {
+          debugPrint('Error creating period notification: $e');
+        }
+
+        // Overdue period notification
+        try {
+          await NotificationService.schedulePeriodNotifications(
+            scheduledTime: predictionResult.estimatedStartDate,
+            areEnabled: periodOverdueNotificationsEnabled,
+            daysAfter: periodOverdueNotificationDays,
+            notificationTime: periodOverdueNotificationTime,
+            title: l10n.notification_periodOverdueTitle,
+            body: l10n.notification_periodOverdueBody(periodOverdueNotificationDays),
+            notificationID: periodOverdueNotificationId,
+          );
+        } catch (e) {
+          debugPrint('Error creating period overdue notification: $e');
+        }
+      }
     }
     
-    if (!mounted) return;
-
     final isPeriodOngoing = periods.isNotEmpty && DateUtils.isSameDay(periods.first.endDate, DateTime.now());
     FabState currentState;
     if (!isPeriodOngoing) {
@@ -128,22 +170,49 @@ class LogsScreenState extends State<LogsScreen> {
     }
     widget.onFabStateChange(currentState);
 
+    int daysUntilDueForCircle = predictionResult?.daysUntilDue ?? 0; 
+		int circleMaxValue = predictionResult?.averageCycleLength ?? 28;
+		int circleCurrentValue = daysUntilDueForCircle.clamp(0, circleMaxValue); 
+
     setState(() {
       _isLoading = false;
       _periodLogEntries = periodDays;
       _periodEntries = periods;
       _predictionResult = predictionResult;
       _selectedView = selectedView;
+      _circleCurrentValue = circleCurrentValue;
+      _circleMaxValue = circleMaxValue;
     });
+
+    await _watchSyncService.sendCircleData(
+      circleMaxValue: _circleMaxValue,
+      circleCurrentValue: _circleCurrentValue,
+    );
   }
 
-  void _handleSaveLog(PeriodDay updatedLog) {
-    periodsRepo.updatePeriodLog(updatedLog);
-    Navigator.of(context).pop();
+  /// Creates a new log entry.
+  Future<void> createNewLog(DateTime selectedDate) async {
+    final bool wasLogSuccessful = await PeriodLoggerService.showAndLogPeriod(context, selectedDate);
+
+    if (wasLogSuccessful && mounted) {
+      _refreshPeriodLogs();
+    }
+  }
+
+  /// updates an existing log entry.
+  Future<void> _updateExistingLog(PeriodDay updatedLog) async {
+    await periodsRepo.updatePeriodLog(updatedLog);
+
+    if (mounted) {
+      Navigator.of(context).pop();
+    }
+
     _refreshPeriodLogs();
   }
 
-	Future<void> _deletePeriodEntry(int id) async {
+  /// Deletes a log entry.
+	Future<void> _deleteExistingLog(int? id) async {
+    if (id == null) return;
 		await periodsRepo.deletePeriodLog(id);
 		_refreshPeriodLogs();
 	}
@@ -151,10 +220,6 @@ class LogsScreenState extends State<LogsScreen> {
 	@override
 	Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context)!;
-
-		int daysUntilDueForCircle = _predictionResult?.daysUntilDue ?? 0; 
-		int circleMaxValue = _predictionResult?.averageCycleLength ?? 28;
-		int circleCurrentValue = daysUntilDueForCircle.clamp(0, circleMaxValue); 
 
 		String predictionText = '';
 		if (_isLoading) {
@@ -178,8 +243,8 @@ class LogsScreenState extends State<LogsScreen> {
       children: [
         const SizedBox(height: 100),
         BasicProgressCircle(
-          currentValue: circleCurrentValue,
-          maxValue: circleMaxValue,
+          currentValue: _circleCurrentValue,
+          maxValue: _circleMaxValue,
           circleSize: 220,
           strokeWidth: 20,
           progressColor: const Color.fromARGB(255, 255, 118, 118),
@@ -202,9 +267,8 @@ class LogsScreenState extends State<LogsScreen> {
           periodLogEntries: _periodLogEntries,
           periodEntries: _periodEntries,
           isLoading: _isLoading,
-          onDelete: _deletePeriodEntry,
-          onSave: _handleSaveLog,
-          onLogRequested: handleLogPeriod,
+          onLogRequested: createNewLog,
+          onLogTapped: _showDetailsBottomSheet,
         ),
       ],
     );
